requirements.txt
python-dateutil==2.9.0.post0
pydantic==2.9.2
google-api-python-client==2.143.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
rich==13.9.2
requests==2.32.3
